move services directory
